+++ components/Card/sub-components/index.js
@@ -1,0 +1,8 @@
+import Thumbnail from './Thumbnail';
+import Container from './Container';
+import { HeaderText, Title, SupportText } from './HeaderText';
+import Media from './Media';
+import Button from './Button';
+import Divider from './Divider';
+
+export { Container, Thumbnail, HeaderText, Title, SupportText, Media, Button, Divider };