+++ CHANGELOG.md
@@ -1,22 +1,39 @@
+## v0.5.3 - May/2018
+### Added
+Url loader
+
+## v0.5.2 - May/2018
+### Changed
+Exporting components
+
+## v0.5.0 - May/2018
+### Added
+New component:
+- Card component
+
+## v0.4.0 - May/2018
+### Added
+New component:
+- Tooltip component
+
 ## v0.3.4 - May/2018
 ### Changed
 Add TextMask Component
 
 ## v0.3.3 - May/2018
-### Changed
+### Added
 Add default font-family (typography)
 
 ## v0.3.2 - May/2018
-### Changed
+### Added
 Add CLI documentation
 
 ## v0.3.1 - May/2018
 ### Changed
-New components:
 - Version fix
 
 ## v0.3.0 - May/2018
-### Changed
+### Added
 New components:
 - Image
 - Logo